% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Dataset}
\format{
A dataframe with 15 rows and two variables
\describe{
   \item{DISTANCE}{Distance of the fire from the station in miles}
   \item{DAMAGE}{X 1000 Dollars in damage caused by the fire}
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
A dataset collected regarding home damage caused by fires.
}
\keyword{datasets}
