% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{My Confidence Interval Function}
\usage{
myci(x, conf.level = 0.95)
}
\arguments{
\item{x}{The data for the function to find the confidence interval}

\item{conf.level}{The level of confidence set for the desired interval}
}
\value{
A confidence interval corresponding to the confidence level set in the parameters, 0.95 by default.
}
\description{
My Confidence Interval Function
}
\examples{
myci(x=c(0,1,3,2,4,5,1,3,12,3,4), conf.level=0.8)
}
