% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{My Central Limit Theorem Function for the Poisson Distribution}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{Sample size}

\item{iter}{Number of iterations to be performed}

\item{lambda}{Mean number of events within a time or space (as described in a Poisson)}

\item{...}{Extra modifications to the histogram of w}
}
\value{
A histogram of the sample means, a barplot of the sampled y, and a probability function for given Poisson parameters
}
\description{
My Central Limit Theorem Function for the Poisson Distribution
}
\examples{
mycltp(n=20,iter=10000,lambda=10)
}
