% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{My Maximum Likelihood Estimator Function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{The log function of the d-stem distribution with x=x and param=param, with select sizes}

\item{x}{The collection of estimated parameters from the data}

\item{param}{The range of parameter values for the maximum likelihood estimate to choose from}

\item{...}{Other parameters for the resulting plot}
}
\value{
A plot of the log function with the Maximum Likelihood Estimate labeled and a list with the index, the parameter of each index, the sum of parameters at each index, and the slope values.
}
\description{
My Maximum Likelihood Estimator Function
}
