% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Distribution Probability Plot}
\usage{
myncurve(a, mu, sigma, col = "red")
}
\arguments{
\item{a}{Upper limit of probability}

\item{mu}{Mean}

\item{sigma}{Standard Deviation}

\item{col}{Color of Shaded Probability}
}
\value{
A plot with the shaded area up to a and P(Y<=a)
}
\description{
Normal Distribution Probability Plot
}
\examples{
myncurve(a=6, mu=10, sigma=4, col="red")
}
