% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sim.R
\name{sample_sim}
\alias{sample_sim}
\title{Sample Simulations}
\usage{
sample_sim(x, n, iter = 10, time = 0.5)
}
\arguments{
\item{x}{Number of items sampled}

\item{n}{Sample size}

\item{iter}{Number of iterations}

\item{time}{Time between each iteration}
}
\value{
A plot of proportions of x selected in each individual iteration
}
\description{
Sample Simulations
}
\examples{
sample_sim(x=10, n=100, iter=10, time=0.5)
}
